Keep main test methods inside the test case sub-class
